+++ docs_src/caching/settings.py
@@ -1,0 +1,6 @@
+from esmerald import EsmeraldAPISettings
+from esmerald.caches.redis import RedisCache
+
+
+class CustomSettings(EsmeraldAPISettings):
+    cache_backend = RedisCache(redis_url="redis://localhost:6379")