+++ esmerald/conf/global_settings.py
@@ -6,7 +6,8 @@ from pydantic import AnyUrl
 from pydantic_settings import BaseSettings, SettingsConfigDict
 from typing_extensions import Annotated, Doc
 
-from esmerald import __version__
+from esmerald import __version__  # noqa
+from esmerald.caches.memory import InMemoryCache
 from esmerald.conf.enums import EnvironmentType
 from esmerald.config import CORSConfig, CSRFConfig, OpenAPIConfig, SessionConfig, StaticFilesConfig
 from esmerald.config.asyncexit import AsyncExitConfig
@@ -16,6 +17,7 @@ from esmerald.interceptors.types import Interceptor
 from esmerald.openapi.schemas.v3_1_0 import Contact, License, SecurityScheme
 from esmerald.permissions.types import Permission
 from esmerald.pluggables import Extension, Pluggable
+from esmerald.protocols.cache import CacheBackend
 from esmerald.routing import gateways
 from esmerald.types import (
     APIGateHandler,
@@ -33,7 +35,44 @@ if TYPE_CHECKING:
     from esmerald.types import TemplateConfig  # pragma: no cover
 
 
-class EsmeraldAPISettings(BaseSettings):
+class CacheBackendSettings(BaseSettings):
+    """
+    `CacheBackendSettings` settings object. The main entry-point for any settings
+    used by **any** Esmerald application.
+
+    Esmerald uses these settings to set the `@caches` decorator and use it across
+    the application.
+    """
+
+    model_config = SettingsConfigDict(
+        extra="allow",
+        ignored_types=(cached_property,),
+        arbitrary_types_allowed=True,
+    )
+
+    cache_backend: Annotated[
+        Optional["CacheBackend"],
+        Doc(
+            """
+            Defines the cache backend to be used for caching operations within the application.
+            By default, an in-memory cache is used, but this can be replaced with other
+            implementations such as Redis or Memcached.
+
+            The cache backend should implement the necessary methods for storing, retrieving,
+            and invalidating cached data.
+
+            Read more about this in the official [Esmerald documentation](https://esmerald.dev/caching/).
+
+            !!! Tip
+                For distributed applications, consider using an external caching backend
+                like Redis instead of the default in-memory cache.
+            """
+        ),
+    ] = InMemoryCache()
+    cache_default_ttl: Annotated[int, Doc("Default time-to-live (TTL) for cached items.")] = 300
+
+
+class EsmeraldAPISettings(CacheBackendSettings):
     """
     `EsmeraldAPISettings` settings object. The main entry-point for any settings
     used by **any** Esmerald application.
@@ -737,12 +776,11 @@ class EsmeraldAPISettings(BaseSettings):
         ),
     ] = "https://esmerald.dev/statics/images/favicon.ico"
 
-    # Model configuration
-    model_config = SettingsConfigDict(extra="allow", ignored_types=(cached_property,))
-
     # Shell configuration
     ipython_args: List[str] = ["--no-banner"]
     ptpython_config_file: str = "~/.config/ptpython/config.py"
+
+    ignore_reload: Annotated[bool, Doc("""Ignore the reload settings.""")] = False
 
     @property
     def reload(self) -> bool: